+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- Fix missing `"types"` field in `package.json`. Types should now be correctly
+  exported.
+
 ## [2.1.0] - 2023-05-10
 
 ### Added